+++ src/panel/device_list/module.html
