+++ src/panel/device_list/plugin.json
